ch: Move mock_endpoints folder -correct mock endpoints folder placement -Amend tests to reflect new folder directory [Delivers #155720290]
