+++ __tests__/TimelineChat.test.jsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import shallowToJSON from 'enzyme-to-json';
 
 import TimelineChat from '../src/Components/TimelineChat/TimelineChat.Component';
-import { testIncident } from '../src/mock_endpoints/mockData';
+import { testIncident } from '../mock_endpoints/mockData';
 
 describe('Timeline Chat component', () => {
   it('should have all the Timeline Chat content', () => {
