+++ __tests__/TimelineNotes.test.jsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import shallowToJSON from 'enzyme-to-json';
 
 import TimelineNotes from '../src/Components/TimelineNotes/TimelineNotes.Component';
-import { testIncident } from '../src/mock_endpoints/mockData';
+import { testIncident } from '../mock_endpoints/mockData';
 
 describe('Timeline Notes component', () => {
   it('should have all the Timeline Notes content', () => {
